Refactored Model class, so other functions also needed refactoring
