+++ app/src/main/java/com/example/splendormobilegame/websocket/UserReaction.java
@@ -1,15 +1,14 @@
 package com.example.splendormobilegame.websocket;
 
+import com.example.splendormobilegame.model.Model;
 import com.github.splendor_mobile_game.websocket.communication.ServerMessage;
 import com.github.splendor_mobile_game.websocket.communication.UserMessage;
 
 public abstract class UserReaction {
     protected ServerMessage serverMessage;
-    protected Model model;
 
-    public UserReaction(ServerMessage serverMessage, Model model) {
+    public UserReaction(ServerMessage serverMessage) {
         this.serverMessage = serverMessage;
-        this.model = model;
     }
 
     public abstract UserMessage react();
